pyproject.toml
[project]
name = "comet"
version = "0.1.0"
description = "Stremio's fastest torrent/debrid search add-on."
authors = [
    { name = "Goldy", email = "153996346+g0ldyy@users.noreply.github.com" }
]
requires-python = ">=3.11"
dependencies = [
    "aiohttp",
    "aiosqlite",
    "asyncio",
    "asyncpg",
    "bencode-py",
    "curl-cffi",
    "databases",
    "demagnetize",
    "fastapi",
    "gunicorn",
    "jinja2",
    "loguru",
    "mediaflow-proxy",
    "orjson",
    "pydantic-settings",
    "python-multipart",
    "rank-torrent-name",
    "uvicorn",
    "python-libtorrent",
]
